requirements.txt
streamlit>=1.35.0
pandas==2.2.1
openpyxl==3.1.2
requests==2.31.0
xlrd==2.0.1
fpdf2>=2.7.9
psycopg2-binary>=2.9.0
